SmartRecruit/requirements.txt
# LLM Models
llama-cpp-python==0.2.90
transformers==4.45.2
torch==2.1.0
accelerate==0.27.2

# NLP & ML
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# Vector Storage
chromadb==0.4.22
faiss-cpu==1.7.4

# JSON Processing
jsonschema==4.20.0

# API & Web
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3

# Bias Detection & Visualization
scipy==1.11.4
plotly==5.18.0
dash==2.14.2

# Utils
python-dotenv==1.0.0
tqdm==4.66.1
redis==5.0.1
celery==5.3.4